move stream response to utils folder
