+++ pages/utils/getStreamResponse.js
